previous commit caused errors with socket.io.  Now, these issues are resolved, and we can communicate over sockets while still writing to the database. I removed the post requests to the server for messages, opting to switch those to sockets.  Message db writes are now done on the message socket emission
